refactor BSP server setup in separate function
[ci skip-rust]

[ci skip-build-wheels]

move to testutil

[ci skip-rust]

[ci skip-build-wheels]
